Remove unnecessary 'type' key
