+++ src/scss/_margarita.scss
@@ -1,4 +1,6 @@
-@import '../css/variables.css';
+// Auto-generated CSS custom properties
+@import '../css/tokens.css';
+
 @import 'margarita-tokens';
 @import 'normalize';
 @import 'transitions/slide-fade';
