+++ src/style-dictionary/styleDictionaryTokens.js
@@ -1,12 +1,12 @@
-const addTokenType = require('./addTokenType')
+const transformToStyleDictionary = require('./transformToStyleDictionary')
 const color = require('../tokens/color')
 const spacing = require('../tokens/spacing')
 const shadow = require('../tokens/shadow')
 const { breakpoints } = require('../tokens/breakpoint')
 
 module.exports = {
-  color: addTokenType(color, 'color'),
-  spacing: addTokenType(spacing, 'spacing'),
-  shadow: addTokenType(shadow, 'shadow'),
-  breakpoint: addTokenType(breakpoints, 'breakpoint'),
+  color: transformToStyleDictionary(color),
+  spacing: transformToStyleDictionary(spacing),
+  shadow: transformToStyleDictionary(shadow),
+  breakpoint: transformToStyleDictionary(breakpoints),
 }
