+++ src/style-dictionary/transformToStyleDictionary.js
@@ -1,17 +1,17 @@
 const merge = require('lodash.merge')
 const isPlainObject = require('lodash.isplainobject')
 
-const addTokenType = (token, type) =>
+const addTokenType = (token) =>
   Object.entries(token).reduce((acc, [key, keyVal]) => {
     if (isPlainObject(keyVal)) {
       return merge(acc, {
-        [key]: addTokenType(keyVal, type),
+        [key]: addTokenType(keyVal),
       })
     }
 
     return {
       ...acc,
-      [key]: { value: keyVal, type },
+      [key]: { value: keyVal },
     }
   }, {})
